Use url hash instead of path
The url path is used by github-pages to decide which repository to
serve. This causes a problem as the code is being interpreted by the
server, where it should just be passed to the client. The server ignores
the url hash, so it is a good candidate for our use case.